added double computation to kelondroMapObjects
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@3316 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ htroot/Network.java
@@ -95,7 +95,7 @@ public class Network {
 
 //          final boolean complete = ((post == null) ? false : post.get("links", "false").equals("true"));
             final long otherppm = yacyCore.seedDB.countActivePPM();
-            final long otherqpm = yacyCore.seedDB.countActiveQPM();
+            final double otherqpm = yacyCore.seedDB.countActiveQPM();
             long myppm = 0;
             double myqph = 0d;
